Title: 1시간 뒤 시간 출력, Time: -1ms, Memory: -1MB, Status: Compile Error - Codetree

+++ 250408/1시간 뒤 시간 출력/print-one-hour-later.java
@@ -9,6 +9,6 @@ public class Main {
 
         String[] str = c.split(":");
 
-        System.out.print((str[0] + 1) + ":" + str[1]);
+        System.out.print(((int)str[0] + 1) + ":" + str[1]);
     }
 }
